Started with OPC UA support, extended HTTP server ======================================================
System Information:
Engine Version: Molecule Ver 1.7.6.20220607
Library Version:  Ver 2.5.0.20220601

Hosts:
Interactive Console Ver 1.0.5.20220607
Pipe Child Ver 1.1.2.20220607
Headless Host Ver 1.0.0.20220607

Storages:
File System Storage Ver 1.2.0.20220607

Firmware:
Windows Ver 1.2.1.20220429
Linux Ver 1.1.0.20220601

Tested Compilers:
MSVC 14.29.30145
GCC 9.4.0

Tested OSes:
Linux Ver:5.13.0-44-generic [x86_64]
Windows Workstation 10.0.19044 [x64]

+++ lib/rx_rt_data.cpp
@@ -385,7 +385,7 @@ rx_simple_value runtime_values_data::get_value (const string_type& path) const
 				if (array_idx < 0 && std::holds_alternative<value_type>(val_it->second))
 					return std::get<value_type>(val_it->second);
 				else if ((int)array_idx >= 0 && std::holds_alternative<array_value_type>(val_it->second)
-					&& array_idx < std::get<array_value_type>(val_it->second).size())
+					&& array_idx < (int)std::get<array_value_type>(val_it->second).size())
 					return std::get<array_value_type>(val_it->second)[idx];
 
 			}
